feat: ⚡️ improve web performance & add (experimental) container queries
