+++ packages/core/src/css/breakpoint-context.ts
@@ -1,0 +1,5 @@
+import { createContext, useContext } from 'react'
+
+export const BreakpointIndexContext = createContext<number>(0)
+
+export const useBreakpointIndex = () => useContext(BreakpointIndexContext)
