+++ packages/core/src/index.ts
@@ -1,12 +1,13 @@
 export * from './components'
 export * from './css/create-themed-component'
 export * from './provider'
-export { setBreakpoints } from './css/breakpoints'
 export * from './css/scales'
-export * from './css/use-breakpoint-index'
+export { useResponsiveValue } from './css/use-responsive-value'
 export { Styles, css } from './css'
 export type { DripsyFinalTheme as Theme } from './declarations'
 export { styled } from './css/styled'
+export { useBreakpointIndex } from './css/breakpoint-context'
+export { getBreakpointIndex } from './css/get-breakpoint-index'
 
 export { useDripsyTheme } from './use-dripsy-theme'
