+++ packages/core/src/use-sx.ts
@@ -2,7 +2,7 @@ import { useCallback, useRef } from 'react'
 import { useDripsyTheme } from './use-dripsy-theme'
 import { SxProp } from './css/types'
 import { css } from './css'
-import { useBreakpointIndex } from './css/use-breakpoint-index'
+import { useBreakpointIndex } from './css/breakpoint-context'
 import stableHash from 'stable-hash'
 import { DripsyFinalTheme } from './declarations'
 import { StyleProp } from 'react-native'
